Fix vulkan timeline semaphores example (#85)
* VulkanTimelineSemaphores: fix descriptor pool size

This patch avoid a VK_ERROR_OUT_OF_POOL_MEMORY error when the swapchain
length is greater than 3.

* Vulkan: fix requested swapchain length

Swapchain length attribute is used to fill minImageCount parameter in
the vkSwapchainCreateInfoKHR structure. By doing max(3,
surfaceCap.minImageCount) we're always choosing the wrong length. Switch
max instruction to min to always choose between 3 (the choice of the
app) or the surface capability if it doesn't support triple buffering.